requirements.txt
aio-pika==4.9.1
marshmallow==2.16.3
nexmo==2.3.0
responder==1.1.2
starlette==0.8.8
python-arango==4.2.1
tenacity==5.0.2
